# perf(lambda-at-edge): fix rollup bundling for dynamic imports perf