--- a/packages/libs/lambda-at-edge/package.json
+++ b/packages/libs/lambda-at-edge/package.json
@@ -51,7 +51,7 @@
     "@types/send": "^0.14.5",
     "@types/sharp": "^0.26.1",
     "fetch-mock-jest": "^1.5.1",
-    "rollup": "^2.26.6",
+    "rollup": "^2.46.0",
     "rollup-plugin-node-externals": "^2.2.0",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-typescript2": "^0.27.2",