--- a/packages/libs/lambda-at-edge/src/build.ts
+++ b/packages/libs/lambda-at-edge/src/build.ts
@@ -234,10 +234,12 @@ class Builder {
     destination: string,
     shouldMinify: boolean
   ) {
-    const source = require.resolve(
-      `@sls-next/lambda-at-edge/dist/${handlerType}${
-        shouldMinify ? ".min" : ""
-      }.js`
+    const source = path.dirname(
+      require.resolve(
+        `@sls-next/lambda-at-edge/dist/${handlerType}${
+          shouldMinify ? "minified" : "standard"
+        }`
+      )
     );
 
     await fse.copy(source, destination);
@@ -295,7 +297,7 @@ class Builder {
       ...copyTraces,
       this.processAndCopyHandler(
         "default-handler",
-        join(this.outputDir, DEFAULT_LAMBDA_CODE_DIR, "index.js"),
+        join(this.outputDir, DEFAULT_LAMBDA_CODE_DIR),
         !!this.buildOptions.minifyHandlers
       ),
       this.buildOptions?.handler
@@ -388,7 +390,7 @@ class Builder {
       ...copyTraces,
       this.processAndCopyHandler(
         "api-handler",
-        join(this.outputDir, API_LAMBDA_CODE_DIR, "index.js"),
+        join(this.outputDir, API_LAMBDA_CODE_DIR),
         !!this.buildOptions.minifyHandlers
       ),
       this.buildOptions?.handler
@@ -422,7 +424,7 @@ class Builder {
     return Promise.all([
       this.processAndCopyHandler(
         "image-handler",
-        join(this.outputDir, IMAGE_LAMBDA_CODE_DIR, "index.js"),
+        join(this.outputDir, IMAGE_LAMBDA_CODE_DIR),
         !!this.buildOptions.minifyHandlers
       ),
       this.buildOptions?.handler