--- a/packages/libs/lambda-at-edge/src/s3-client.ts
+++ b/packages/libs/lambda-at-edge/src/s3-client.ts
@@ -1,5 +1,0 @@
-import { GetObjectCommand } from "@aws-sdk/client-s3/commands/GetObjectCommand";
-import { PutObjectCommand } from "@aws-sdk/client-s3/commands/PutObjectCommand";
-import { S3Client } from "@aws-sdk/client-s3/S3Client";
-
-export { GetObjectCommand, PutObjectCommand, S3Client };