… packages/libs/lambda-at-edge/yarn.lock