Note: testing rollup config to properly bundle dynamic imports like s3client which is not always needed. The cost is an additional wrapper on top of main code but this seems expected optimization from Rollup.js (preloading and parsing optimization). 

Previously, inlining dynamic imports into one file was still causing that code to be at least partially preloaded outside of the handler, e.g initializing some S3/AWS SDK related objects even if the S3Client itself was only initialized when needed. So it caused higher cold starts due to this for all paths. With this change, we split into multiple chunks. Total there should be a few chunks:

* Entry chunk (wrapper of ~20 lines on top of other code)
* Core chunk (all of the routing code bundled into one file)
* Dynamic chunks (e.g S3 client, GetObjectCommand, PutObjectCommand) which is heavy code only needed on fallback paths

Total size of all chunks remains around the same but slightly higher at 657 kB vs. 651 kB before. However, Lambda does not need to read (and preload some parts of it) ~166 kB of code for S3 unless it's needed. There is one extra file I/O due to the wrapper requiring the core code but from tests, it appears to be negligible as ~25 ms is saved in init time from my tests (195 -> 170 ms with the 512 MB Lambda function)
In the future, we can do further optimizations like getting rid of complexity of having the origin response handler, and even using Rollup again at build/deploy time to further get rid of code paths that are not used (e.g if user has no rewrites, we can get rid of that code)