@jvarho 2021-04-29
Yes, this looks better than my effort :+1:

I can confirm this reproduces the improvement I saw in cold starts. Here are five-run averages for before  and after this branch. I'm using 128 MB memory setting and a redirect to get the worst case.

Before:

Max Memory Used: 71 MB
Init Duration: 201.8 +/- 8.3 ms
Duration: 30.2 +/- 5.3 ms

After:

Max Memory Used: 67 MB
Init Duration: 161.2 +/- 6.9 ms
Duration: 31.3 +/- 6.7 ms

About 20% lower init duration and a bit less memory use.

I have not tested the fallback time, because it's dominated by S3 latency and rendering.